+++ ocr/src/main/java/com/ing/software/ocr/OcrObjects/TicketSchemes/TicketSchemeIT_PC.java
@@ -1,6 +1,7 @@
 package com.ing.software.ocr.OcrObjects.TicketSchemes;
 
 import android.support.annotation.NonNull;
+import android.util.Pair;
 
 import com.ing.software.common.Scored;
 
@@ -8,6 +9,7 @@ import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import com.annimon.stream.Stream;
+import com.ing.software.ocr.OcrObjects.OcrText;
 import com.ing.software.ocr.OcrUtils;
 
 /**
@@ -18,13 +20,16 @@ import com.ing.software.ocr.OcrUtils;
 public class TicketSchemeIT_PC implements TicketScheme{
 
     private String tag = "IT_PC";
-    private List<BigDecimal> products = new ArrayList<>();
+    private List<Pair<OcrText, BigDecimal>> products = new ArrayList<>();
     private BigDecimal total;
     private BigDecimal cash;
+    private BigDecimal bestAmount;
+    private boolean acceptedList = false;
+    private final int THREE_VALUES = 80; //MAX SCORE
 
-    public TicketSchemeIT_PC(BigDecimal total, @NonNull List<BigDecimal> aboveTotal, @NonNull List<BigDecimal> belowTotal) {
+    public TicketSchemeIT_PC(BigDecimal total, @NonNull List<Pair<OcrText, BigDecimal>> aboveTotal, @NonNull List<BigDecimal> belowTotal) {
         this.total = total;
-        this.products = new ArrayList<>(aboveTotal); //Necessary or in psc last index is removed
+        this.products = new ArrayList<>(aboveTotal);
         if (aboveTotal.isEmpty())
             products = null;
         if (!belowTotal.isEmpty()) {
@@ -33,13 +38,28 @@ public class TicketSchemeIT_PC implements TicketScheme{
     }
 
     @Override
-    public Scored<BigDecimal> getBestAmount(boolean strict) {
-        return strict ? strictBestAmount() : looseBestAmount();
+    public BigDecimal getBestAmount() {
+        return bestAmount;
+    }
+
+    @Override
+    public double getAmountScore(boolean strict) {
+        Scored<BigDecimal> tempAmount = strict ? strictBestAmount() : looseBestAmount();
+        if (tempAmount != null) {
+            bestAmount = tempAmount.obj();
+            return tempAmount.getScore();
+        } else
+            return -1;
     }
 
     @Override
     public String toString() {
         return tag;
+    }
+
+    @Override
+    public List<Pair<OcrText, BigDecimal>> getPricesList() {
+        return acceptedList ? products : null;
     }
 
     /**
@@ -48,12 +68,15 @@ public class TicketSchemeIT_PC implements TicketScheme{
     private Scored<BigDecimal> strictBestAmount() {
         if (products != null && total != null && cash != null) {
             BigDecimal productsSum = Stream.of(products)
+                    .map(product -> product.second)
                     .reduce(BigDecimal.ZERO, BigDecimal::add);
             OcrUtils.log(3, "TicketScheme_" + tag, "productsSum is: " + productsSum);
             OcrUtils.log(3, "TicketScheme_" + tag, "total is: " + total);
             OcrUtils.log(3, "TicketScheme_" + tag, "cash is: " + cash);
-            if (productsSum.compareTo(total) == 0 && cash.compareTo(total) == 0)
-                return new Scored<>(100, total);
+            if (productsSum.compareTo(total) == 0 && cash.compareTo(total) == 0) {
+                acceptedList = true;
+                return new Scored<>(THREE_VALUES, total);
+            }
         }
         return null;
     }
@@ -62,18 +85,19 @@ public class TicketSchemeIT_PC implements TicketScheme{
      * @return best amount according to arbitrary decisions
      */
     private Scored<BigDecimal> looseBestAmount() {
-        int THREE_VALUES = 80;
         int TWO_VALUES_AMOUNT = 50;
         int TWO_VALUES = 30;
         int NO_MATCH = 1;
         if (total != null) {
             if (products != null && cash != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 boolean cashPrices = cash.compareTo(productsSum) == 0;
                 boolean cashAmount = cash.compareTo(total) == 0;
                 boolean pricesAmount = productsSum.compareTo(total) == 0;
                 if (cashPrices || pricesAmount) {
+                    acceptedList = true;
                     if (cashAmount) {
                         return new Scored<>(THREE_VALUES, total);
                     } else if (pricesAmount) {
@@ -86,8 +110,10 @@ public class TicketSchemeIT_PC implements TicketScheme{
                 }
             } else if (products != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 if (productsSum.compareTo(total) == 0) {
+                    acceptedList = true;
                     return new Scored<>(TWO_VALUES_AMOUNT, total);
                 } else {
                     return new Scored<>(NO_MATCH, total);
@@ -105,8 +131,10 @@ public class TicketSchemeIT_PC implements TicketScheme{
             //no total
             if (products != null && cash != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 if (cash.compareTo(productsSum) == 0) {
+                    acceptedList = true;
                     return new Scored<>(TWO_VALUES, cash);
                 }
             }