+++ ocr/src/main/java/com/ing/software/ocr/OcrObjects/TicketSchemes/TicketSchemeIT_PCC.java
@@ -1,9 +1,11 @@
 package com.ing.software.ocr.OcrObjects.TicketSchemes;
 
 import android.support.annotation.NonNull;
+import android.util.Pair;
 
 import com.annimon.stream.Stream;
 import com.ing.software.common.Scored;
+import com.ing.software.ocr.OcrObjects.OcrText;
 import com.ing.software.ocr.OcrUtils;
 
 import java.math.BigDecimal;
@@ -18,12 +20,15 @@ import java.util.List;
 public class TicketSchemeIT_PCC implements TicketScheme{
 
     private String tag = "IT_PCC";
-    private List<BigDecimal> products = new ArrayList<>();
+    private List<Pair<OcrText, BigDecimal>> products = new ArrayList<>();
     private BigDecimal total;
     private BigDecimal cash;
     private BigDecimal change;
+    private BigDecimal bestAmount;
+    private boolean acceptedList = false;
+    private final int THREE_VALUES = 80;
 
-    public TicketSchemeIT_PCC(BigDecimal total, @NonNull List<BigDecimal> aboveTotal, @NonNull List<BigDecimal> belowTotal) {
+    public TicketSchemeIT_PCC(BigDecimal total, @NonNull List<Pair<OcrText, BigDecimal>> aboveTotal, @NonNull List<BigDecimal> belowTotal) {
         this.total = total;
         products = new ArrayList<>(aboveTotal);
         if (aboveTotal.isEmpty())
@@ -36,13 +41,28 @@ public class TicketSchemeIT_PCC implements TicketScheme{
     }
 
     @Override
-    public Scored<BigDecimal> getBestAmount(boolean strict) {
-        return strict ? strictBestAmount() : looseBestAmount();
+    public BigDecimal getBestAmount() {
+        return bestAmount;
+    }
+
+    @Override
+    public double getAmountScore(boolean strict) {
+        Scored<BigDecimal> tempAmount = strict ? strictBestAmount() : looseBestAmount();
+        if (tempAmount != null) {
+            bestAmount = tempAmount.obj();
+            return tempAmount.getScore();
+        } else
+            return -1;
     }
 
     @Override
     public String toString() {
         return tag;
+    }
+
+    @Override
+    public List<Pair<OcrText, BigDecimal>> getPricesList() {
+        return acceptedList ? products : null;
     }
 
     /**
@@ -52,12 +72,15 @@ public class TicketSchemeIT_PCC implements TicketScheme{
         if (products != null && total != null && cash != null && change != null) {
             BigDecimal normCash = cash.subtract(change);
             BigDecimal productsSum = Stream.of(products)
+                    .map(product -> product.second)
                     .reduce(BigDecimal.ZERO, BigDecimal::add);
             OcrUtils.log(3, "TicketScheme_" + tag, "productsSum is: " + productsSum);
             OcrUtils.log(3, "TicketScheme_" + tag, "total is: " + total);
             OcrUtils.log(3, "TicketScheme_" + tag, "cash is: " + normCash);
-            if (productsSum.compareTo(total) == 0 && normCash.compareTo(total) == 0)
-                return new Scored<>(100, total);
+            if (productsSum.compareTo(total) == 0 && normCash.compareTo(total) == 0){
+                acceptedList = true;
+                return new Scored<>(THREE_VALUES, total);
+            }
         }
         return null;
     }
@@ -66,19 +89,20 @@ public class TicketSchemeIT_PCC implements TicketScheme{
      * @return best amount according to arbitrary decisions
      */
     private Scored<BigDecimal> looseBestAmount() {
-        int THREE_VALUES = 80;
         int TWO_VALUES_AMOUNT = 50;
         int TWO_VALUES = 30;
         int NO_MATCH = 1;
         if (total != null) {
             if (products != null && cash != null && change != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 BigDecimal normCash = cash.subtract(change);
                 boolean cashPrices = normCash.compareTo(productsSum) == 0;
                 boolean cashAmount = normCash.compareTo(total) == 0;
                 boolean pricesAmount = productsSum.compareTo(total) == 0;
                 if (cashPrices || pricesAmount) {
+                    acceptedList = true;
                     if (cashAmount) {
                         return new Scored<>(THREE_VALUES, total);
                     } else if (pricesAmount) {
@@ -91,8 +115,10 @@ public class TicketSchemeIT_PCC implements TicketScheme{
                 }
             } else if (products != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 if (productsSum.compareTo(total) == 0) {
+                    acceptedList = true;
                     return new Scored<>(TWO_VALUES_AMOUNT, total);
                 } else {
                     return new Scored<>(NO_MATCH, total);
@@ -111,9 +137,11 @@ public class TicketSchemeIT_PCC implements TicketScheme{
             //no total
             if (products != null && cash != null && change != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 BigDecimal normCash = cash.subtract(change);
                 if (normCash.compareTo(productsSum) == 0) {
+                    acceptedList = true;
                     return new Scored<>(TWO_VALUES, normCash);
                 }
             }