+++ ocr/src/main/java/com/ing/software/ocr/OcrObjects/TicketSchemes/TicketSchemeIT_PSCC.java
@@ -1,9 +1,11 @@
 package com.ing.software.ocr.OcrObjects.TicketSchemes;
 
 import android.support.annotation.NonNull;
+import android.util.Pair;
 
 import com.annimon.stream.Stream;
 import com.ing.software.common.Scored;
+import com.ing.software.ocr.OcrObjects.OcrText;
 import com.ing.software.ocr.OcrUtils;
 
 import java.math.BigDecimal;
@@ -19,16 +21,20 @@ import java.util.List;
 public class TicketSchemeIT_PSCC implements TicketScheme{
 
     private String tag = "IT_PSCC";
-    private List<BigDecimal> products = new ArrayList<>();
+    private List<Pair<OcrText, BigDecimal>> products = new ArrayList<>();
     private BigDecimal subtotal;
     private BigDecimal total;
     private BigDecimal cash;
     private BigDecimal change;
+    private BigDecimal bestAmount;
+    private boolean acceptedList = false;
+    private final int FOUR_VALUES = 100;
 
-    public TicketSchemeIT_PSCC(BigDecimal total, @NonNull List<BigDecimal> aboveTotal, @NonNull List<BigDecimal> belowTotal) {
+
+    public TicketSchemeIT_PSCC(BigDecimal total, @NonNull List<Pair<OcrText, BigDecimal>> aboveTotal, @NonNull List<BigDecimal> belowTotal) {
         this.total = total;
         if (!aboveTotal.isEmpty()) {
-            this.subtotal = aboveTotal.get(aboveTotal.size() - 1);
+            this.subtotal = aboveTotal.get(aboveTotal.size() - 1).second;
         }
         if (aboveTotal.size() > 1) {
             products = new ArrayList<>(aboveTotal);
@@ -42,13 +48,28 @@ public class TicketSchemeIT_PSCC implements TicketScheme{
     }
 
     @Override
-    public Scored<BigDecimal> getBestAmount(boolean strict) {
-        return strict ? strictBestAmount() : looseBestAmount();
+    public BigDecimal getBestAmount() {
+        return bestAmount;
+    }
+
+    @Override
+    public double getAmountScore(boolean strict) {
+        Scored<BigDecimal> tempAmount = strict ? strictBestAmount() : looseBestAmount();
+        if (tempAmount != null) {
+            bestAmount = tempAmount.obj();
+            return tempAmount.getScore();
+        } else
+            return -1;
     }
 
     @Override
     public String toString() {
         return tag;
+    }
+
+    @Override
+    public List<Pair<OcrText, BigDecimal>> getPricesList() {
+        return acceptedList ? products : null;
     }
 
     /**
@@ -58,13 +79,16 @@ public class TicketSchemeIT_PSCC implements TicketScheme{
         if (products != null && total != null && cash != null && change != null && subtotal != null) {
             BigDecimal normCash = cash.subtract(change);
             BigDecimal productsSum = Stream.of(products)
+                    .map(product -> product.second)
                     .reduce(BigDecimal.ZERO, BigDecimal::add);
             OcrUtils.log(3, "TicketScheme_" + tag, "productsSum is: " + productsSum);
             OcrUtils.log(3, "TicketScheme_" + tag, "total is: " + total);
             OcrUtils.log(3, "TicketScheme_" + tag, "cash is: " + normCash);
             OcrUtils.log(3, "TicketScheme_" + tag, "subtotal is: " + subtotal);
-            if (productsSum.compareTo(total) == 0 && normCash.compareTo(total) == 0 && subtotal.compareTo(total) == 0)
-                return new Scored<>(100, total);
+            if (productsSum.compareTo(total) == 0 && normCash.compareTo(total) == 0 && subtotal.compareTo(total) == 0){
+                acceptedList = true;
+                return new Scored<>(FOUR_VALUES, total);
+            }
         }
         return null;
     }
@@ -73,7 +97,6 @@ public class TicketSchemeIT_PSCC implements TicketScheme{
      * @return best amount according to arbitrary decisions
      */
     private Scored<BigDecimal> looseBestAmount() {
-        int FOUR_VALUES = 100;
         int THREE_VALUES = 65;
         int THREE_VALUES_AMOUNT = 75;
         int TWO_VALUES_AMOUNT = 55;
@@ -82,6 +105,7 @@ public class TicketSchemeIT_PSCC implements TicketScheme{
         if (total != null) {
             if (products != null && cash != null && subtotal != null && change != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 BigDecimal normCash = cash.subtract(change);
                 boolean cashPrices = normCash.compareTo(productsSum) == 0;
@@ -92,33 +116,43 @@ public class TicketSchemeIT_PSCC implements TicketScheme{
                 boolean subtotalPrices = subtotal.compareTo(productsSum) == 0;
                 if (cashPrices && pricesAmount) {
                     if (subtotalAmount) {
+                        acceptedList = true;
                         return new Scored<>(FOUR_VALUES, total);
                     } else {
+                        acceptedList = true;
                         return new Scored<>(THREE_VALUES_AMOUNT, total);
                     }
                 } else if (cashAmount && subtotalAmount) {
                     return new Scored<>(THREE_VALUES_AMOUNT, total);
                 } else if (pricesAmount && subtotalAmount) {
+                    acceptedList = true;
                     return new Scored<>(THREE_VALUES_AMOUNT, total);
                 } else if (cashPrices && subtotalPrices) {
+                    acceptedList = true;
                     return new Scored<>(THREE_VALUES, subtotal);
-                } else if (cashAmount || pricesAmount || subtotalAmount) {
+                } else if (cashAmount || subtotalAmount) {
                     return new Scored<>(TWO_VALUES_AMOUNT, total);
-                } else if (cashPrices || subtotalCash) {
-                    return new Scored<>(TWO_VALUES, normCash);
-                } else if (subtotalPrices) {
+                } else if (pricesAmount) {
+                    acceptedList = true;
+                    return new Scored<>(TWO_VALUES_AMOUNT, total);
+                } else if (cashPrices || subtotalPrices) {
+                    acceptedList = true;
+                    return new Scored<>(TWO_VALUES, productsSum);
+                } else if (subtotalCash) {
                     return new Scored<>(TWO_VALUES, subtotal);
                 } else {
                     return new Scored<>(NO_MATCH, total);
                 }
             } else if (products != null && cash != null && change != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 BigDecimal normCash = cash.subtract(change);
                 boolean cashPrices = normCash.compareTo(productsSum) == 0;
                 boolean cashAmount = normCash.compareTo(total) == 0;
                 boolean pricesAmount = productsSum.compareTo(total) == 0;
                 if (cashPrices || pricesAmount) {
+                    acceptedList = true;
                     if (cashAmount) {
                         return new Scored<>(THREE_VALUES, total);
                     } else if (pricesAmount) {
@@ -131,11 +165,13 @@ public class TicketSchemeIT_PSCC implements TicketScheme{
                 }
             } else if (products != null && subtotal != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 boolean pricesAmount = productsSum.compareTo(total) == 0;
                 boolean subtotalAmount = subtotal.compareTo(total) == 0;
                 boolean subtotalPrices = subtotal.compareTo(productsSum) == 0;
                 if (subtotalPrices || pricesAmount) {
+                    acceptedList = true;
                     if (subtotalAmount) {
                         return new Scored<>(THREE_VALUES, total);
                     } else if (pricesAmount) {
@@ -169,31 +205,38 @@ public class TicketSchemeIT_PSCC implements TicketScheme{
             //no total
             if (products != null && cash != null && subtotal != null && change != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 BigDecimal normCash = cash.subtract(change);
                 boolean cashPrices = normCash.compareTo(productsSum) == 0;
                 boolean subtotalCash = subtotal.compareTo(normCash) == 0;
                 boolean subtotalPrices = subtotal.compareTo(productsSum) == 0;
                 if (cashPrices && subtotalCash) {
+                    acceptedList = true;
                     return new Scored<>(THREE_VALUES, normCash);
-                } else if (cashPrices || subtotalCash) {
-                    return new Scored<>(TWO_VALUES, normCash);
-                } else if (subtotalPrices) {
+                } else if (cashPrices || subtotalPrices) {
+                    acceptedList = true;
+                    return new Scored<>(TWO_VALUES, productsSum);
+                } else if (subtotalCash) {
                     return new Scored<>(TWO_VALUES, subtotal);
                 }
             } else if (products != null && cash != null && change != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 BigDecimal normCash = cash.subtract(change);
                 boolean cashPrices = normCash.compareTo(productsSum) == 0;
                 if (cashPrices) {
+                    acceptedList = true;
                     return new Scored<>(TWO_VALUES, normCash);
                 }
             } else if (products != null && subtotal != null) {
                 BigDecimal productsSum = Stream.of(products)
+                        .map(product -> product.second)
                         .reduce(BigDecimal.ZERO, BigDecimal::add);
                 boolean subtotalPrices = subtotal.compareTo(productsSum) == 0;
                 if (subtotalPrices) {
+                    acceptedList = true;
                     return new Scored<>(TWO_VALUES, subtotal);
                 }
             } else if (cash != null && subtotal != null && change != null) {